requirements.txt
aiofiles==24.1.0
aiogram==3.15.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
Django==5.1.3
django-environ==0.11.2
django-solo==2.4.0
environ==1.0
environs==11.2.1
frozenlist==1.5.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
loguru==0.7.2
magic-filter==1.0.12
marshmallow==3.23.1
multidict==6.1.0
packaging==24.2
propcache==0.2.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyOpenSSL==24.3.0
python-dotenv==1.0.1
redis==5.2.0
requests==2.32.3
service-identity==24.2.0
sqlparse==0.5.2
tomli==2.2.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.18.3
zope.interface==7.2
